refactor(legal): extract resource-side form and make copyright holder required
Split the resource-side pane out of `LegalSettingsComponent` into a
new dedicated `ResourceSideLegalFormComponent`. The child receives
the resolved project via `@Input({ required: true })` and builds its
typed FormGroup with initial values seeded from that input in
ngOnInit — so the form is *born* populated, instead of being
hydrated later by a manual `resetResourceSide()` method that reached
into the sync `ProjectPageService.currentProject` getter.

This fixes a user-visible bug where the form loaded empty and only
painted its values after the user interacted with a field: the
previous ordering could apply values to controls before their
options / DOM binding were ready, and only later interactions
triggered the change detection pass that made the stored values
visible.

The `LegalSettingsComponent` shell is now a slim host that owns the
tab layout, gates the child on `@if (project$ | async; as project)`,
and reloads asset-side data on the child's `(saved)` event. The
parent no longer knows anything about form construction, mat-select
options, or the API save call.

UX changes on the copyright-holder field:

- The field is now `Validators.required` — Save stays disabled until
  a holder is provided.
- The old "pre-fill with `project.longname` and mark dirty on load"
  affordance is gone. Instead the longname is offered as a
  suggestion via the placeholder plus a `matSuffix` icon button
  (`content_paste_go`) that fills the field on a single click. The
  button only renders when the field is empty and `longname` exists.
- Cancel becomes a plain `form.reset(...)` from the captured
  `@Input` values — no service round-trip, no manual reset method.

i18n: adds `holderRequired`, `useLongname` (tooltip) and
`useLongnameAria` (aria-label with `{{name}}` interpolation) and
rewrites `holderLabel` and `holderHelper` in en/de/fr/it to reflect
the new behaviour (the label no longer claims the field *is* the
project name, and the helper describes the empty-defaults-to-longname
fallback rather than the old pre-fill).
