+++ apps/dsp-app/src/assets/i18n/it.json
@@ -15,8 +15,11 @@
       "settings": {
         "resourceSide": "Lato risorsa",
         "assetSide": "Lato asset",
-        "holderLabel": "Titolare dei diritti · nome ufficiale del progetto",
-        "holderHelper": "Il vostro progetto è titolare dei diritti di ogni record di risorsa. Il nome ufficiale è indicato di seguito — modificatelo se necessario, quindi aggiungete l'autore.",
+        "holderLabel": "Titolare dei diritti",
+        "holderHelper": "Titolare dei diritti per ogni risorsa di questo progetto. Se il campo è lasciato vuoto, verrà utilizzato il nome ufficiale del progetto.",
+        "holderRequired": "Il titolare dei diritti è obbligatorio.",
+        "useLongname": "Usa il nome ufficiale del progetto",
+        "useLongnameAria": "Usa {{name}} come titolare dei diritti",
         "licensePlaceholder": "Scegli una licenza Creative Commons",
         "liveWarning": "Questa licenza si applica immediatamente a tutte le risorse del progetto.",
         "save": "Salva le modifiche",
